refactor(web): create mapper utils directory
