+++ appeals/web/src/server/appeals/appeal-details/accordions/common/case-overview.js
@@ -1,4 +1,4 @@
-import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { removeSummaryListActions } from '#lib/mappers/index.js';
 import { isDefined } from '#lib/ts-utilities.js';
 
 /**
