+++ appeals/web/src/server/appeals/appeal-details/site-visit/site-visit.mapper.js
@@ -1,5 +1,5 @@
 import { initialiseAndMapAppealData } from '#lib/mappers/appeal/appeal.mapper.js';
-import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { removeSummaryListActions } from '#lib/mappers/index.js';
 import { appealShortReference } from '#lib/appeals-formatter.js';
 import { padNumberWithZero } from '#lib/string-utilities.js';
 import { preRenderPageComponents } from '#lib/nunjucks-template-builders/page-component-rendering.js';
