+++ appeals/web/src/server/lib/mappers/appellant-case/appeal-type-fpa.mapper.js
@@ -1,4 +1,4 @@
-import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { removeSummaryListActions } from '#lib/mappers/index.js';
 import { generateHASComponents } from './appeal-type-has.mapper.js';
 
 /**
