+++ appeals/web/src/server/lib/mappers/appellant-case/appeal-type-has.mapper.js
@@ -1,4 +1,4 @@
-import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { removeSummaryListActions } from '#lib/mappers/index.js';
 import {
 	documentUploadUrlTemplate,
 	mapDocumentManageUrl
