+++ appeals/web/src/server/lib/mappers/index.js
@@ -1,0 +1,1 @@
+export * from './utils/index.js';
