+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-affects-listed-building-details.js
@@ -1,6 +1,6 @@
 import { permissionNames } from '#environment/permissions.js';
 import * as displayPageFormatter from '#lib/display-page-formatter.js';
-import { mapActionComponent } from '../../permissions.mapper.js';
+import { mapActionComponent } from '../../utils/permissions.mapper.js';
 
 /** @type {import('../lpa-questionnaire.mapper.js').SubMapper} */
 export const mapAffectsListedBuildingDetails = ({
