+++ appeals/web/src/server/lib/mappers/utils/remove-summary-list-actions.js
